+++ src/components/com-nav-item/com-nav-item.tsx
@@ -1,4 +1,4 @@
-import { Component, Host, h } from '@stencil/core';
+import { Component, Host, h, Event, EventEmitter, Listen } from '@stencil/core';
 
 /**
  * The com-nav-item is a subcomponent of `wcs-com-nav`. It represents a list-item wrapper around a link.
@@ -9,7 +9,18 @@ import { Component, Host, h } from '@stencil/core';
     shadow: true
 })
 export class ComNavItem {
-
+    /**
+     * Emitted when a user click on a final navigation action.
+     *
+     * Used by the com-nav component to close the mobile menu overlay when a user click on a final action.
+     */
+    @Event() private wcsClickOnFinalAction: EventEmitter<void>;
+    
+    @Listen('click')
+    onClick() {
+        this.wcsClickOnFinalAction.emit();
+    }
+    
     render() {
         return (
             <Host role="listitem">